Fix float64/float32 mismatch: cast ami_score output to float32, use .assign() for checkpoint

+++ util.py
@@ -29,7 +29,7 @@ def ami_score(predictions, targets, K=3):
     predictions = tf.reshape(resp_indices, shape=(-1,))
     targets = tf.reshape(targets, shape=(-1,))
     amis = adjusted_mutual_info_score(predictions.numpy().ravel(), targets.numpy().ravel())
-    return amis
+    return tf.cast(amis, tf.float32)
